\name{BCEndo-package}
\alias{BCEndo-package}
\docType{package}
\title{
\packageTitle{BCEndo}
}
\description{
\packageDescription{BCEndo}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BCEndo}
\packageIndices{BCEndo}

}
\author{
\packageAuthor{BCEndo}

Maintainer: \packageMaintainer{BCEndo}
}
\references{
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{

}
\examples{

}
